+++ test/message_test.exs
@@ -1,0 +1,4 @@
+defmodule X3m.System.MessageTest do
+  use ExUnit.Case, async: true
+  doctest X3m.System.Message
+end