+++ CHANGELOG.md
@@ -79,8 +79,8 @@ use X3m.System.MessageHandler,
 
 # 0.7.16
 
-- `execute_on_new_aggregate returns `{:ok, -1}`if aggregate returns`:ok`response
-with empty`events`
+- `execute_on_new_aggregate` returns `{:ok, -1}` if aggregate returns `:ok` response
+  with empty `events`
 
 # 0.7.15